refactor: restructure API gateway, centralize schemas, consolidate tests
- Create shared schemas/ directory with canonical Pydantic contracts
  (GPSMessage, BusLifecycleState, BusStatusMessage, CoordinateBounds)
  using snake_case fields with camelCase aliases for JSON compatibility

- Flatten api-gateway: remove unnecessary app/ package nesting,
  merge enhanced health/metrics code directly into main.py

- Consolidate all tests into root tests/ directory
  (moved gateway tests from services/api-gateway/tests/)

- Remove duplicate/conflicting schema definitions
  (GPSPayload vs GPSReading, BusState vs BusLifecycleState)

- Fix broken schemas/__init__.py (had double __all__ assignment)

- Update imports: geo_config and gps models now live in schemas/

- Update README repo structure and startup command

All 28 tests passing (27 unit + 1 integration).
