+++ scripts/models/__init__.py
@@ -1,3 +1,2 @@
 from .settings import Settings
 from .route import Stop, RouteGeometry, RouteSeed
-from .gps import GPSMessage
