--- services/api-gateway/app/__init__.py
@@ -1,1 +1,0 @@
-"""API gateway application package."""
