--- services/api-gateway/app/schemas/__init__.py
@@ -1,12 +1,0 @@
-from .bus_status import BusLifecycleState, BusStatusMessage
-from .gps import GPSPayload
-
-__all__ = [
-    "BusLifecycleState",
-    "BusStatusMessage",
-    "GPSPayload",
-]
-from .bus_status import BusStatusEvent, BusState
-from .gps import GPSReading
-
-__all__ = ["GPSReading", "BusStatusEvent", "BusState"]
